# examples/axum-generated-server/openapi.yaml
openapi: "3.0.3"
info:
  title: Notes API
  version: "1.0.0"
paths:
  /notes:
    get:
      operationId: listNotes
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - name: tag
          in: query
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: A list of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      # Secured: the generated `Api` trait injects `auth: Self::Auth` into
      # `create_note` (and only there) and emits a `type Auth` slot.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewNote"
      responses:
        "201":
          description: The created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Note:
      type: object
      required: [id, text]
      properties:
        id:
          type: string
        text:
          type: string
          minLength: 1
        tag:
          type: string
          minLength: 1
    NewNote:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 280
        tag:
          type: string
          minLength: 1